requirements.txt
requests
python-whois
pyOpenSSL
h8mail